merge french vanilla fixes into main

+++ linuxdoom-1.10/m_misc.c
@@ -392,7 +392,7 @@ void M_LoadDefaults (void)
 			    *defaults[i].location = parm;
 			else
 			    *defaults[i].location =
-				(int) newstring;
+				(long long int) newstring;
 			break;
 		    }
 	    }
